+++ R/read_vcf.R
@@ -65,17 +65,17 @@
 #' sumstats_dt <- read_vcf(path = path)
 #' 
 #' #### Remote file ####
-#' ## Small GWAS
-#' path <- "https://gwas.mrcieu.ac.uk/files/ieu-a-298/ieu-a-298.vcf.gz" 
-#' sumstats_dt2 <- read_vcf(path = path)
+#' ## Small GWAS (0.2Mb)
+#' # path <- "https://gwas.mrcieu.ac.uk/files/ieu-a-298/ieu-a-298.vcf.gz" 
+#' # sumstats_dt2 <- read_vcf(path = path)
 #' 
-#' ## Large GWAS
+#' ## Large GWAS (250Mb)
 #' # path <- "https://gwas.mrcieu.ac.uk/files/ubm-a-2929/ubm-a-2929.vcf.gz"
 #' # sumstats_dt3 <- read_vcf(path = path, nThread=11)
 #' 
-#' ### Additional GWAS
+#' ### Very large GWAS (500Mb)
 #' # path <- "https://gwas.mrcieu.ac.uk/files/ieu-a-1124/ieu-a-1124.vcf.gz"
-#' # sumstats_dt <- read_vcf(path = path, nThread=11)
+#' # sumstats_dt4 <- read_vcf(path = path, nThread=11)
 read_vcf <- function(path, 
                      as_datatable = TRUE,
                      save_path = NULL,
@@ -98,6 +98,8 @@ read_vcf <- function(path,
                                 use_params = use_params,
                                 as_datatable = as_datatable,
                                 sampled_rows = sampled_rows,
+                                
+                                download = download,
                                 vcf_dir = vcf_dir,
                                 download_method = download_method,
                                 force_new = force_new,